requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
enum34==1.1.6
idna==2.7
imutils==0.5.1
ipaddress==1.0.22
numpy==1.15.4
opencv-python==3.4.4.19
pathlib==1.0.1
Pillow==5.3.0
pycparser==2.19
pyzbar==0.1.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
